# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Veritae Auth & Member API
  version: 0.1.0
  description: Veritae 인증(회원가입/로그인/토큰 갱신) 및 회원 정보 조회 API.
servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버
...
paths:
...
  /api/v1/analysis/image:
    post:
      tags: [Analysis]
      operationId: analyzeImage
      summary: 이미지 AI 생성 여부 분석
      description: >-
        업로드한 이미지가 AI로 생성되었을 확률을 점수로 반환한다. 탐지 서버(별도 인프라) 호출이 끝날 때까지
        응답을 기다리는 동기 방식이다 - 분석에 수 초~수십 초가 걸릴 수 있다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 분석할 이미지 파일(jpeg/png/webp).
      responses:
        '200':
          description: 분석 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageAnalysisResponse'
              example:
                id: 11111111-1111-1111-1111-111111111111
                aiDetection:
                  model: spai
                  score: 0.000134
        '400':
          $ref: '#/components/responses/InvalidImageFile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/DetectionServiceUnavailable'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/analysis/audio:
    post:
      tags: [Analysis]
      operationId: analyzeAudio
      summary: 음성 AI 생성 여부 분석
      description: >-
        업로드한 음성이 AI로 생성(합성)되었을 확률과 판독 근거를 반환한다. 탐지 서버(별도 인프라) 호출이 끝날 때까지
        응답을 기다리는 동기 방식이다. 재생 길이가 5분을 넘으면 분석하지 않고 400(errorCode
        AUDIO_TOO_LONG)으로 거부한다 - 앞부분만 잘라서 분석하지 않는다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 분석할 음성 파일(wav/mp3/m4a/aac, 최대 5분/25MB).
      responses:
        '200':
          description: 분석 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudioAnalysisResponse'
              example:
                id: 22222222-2222-2222-2222-222222222222
                aiDetection:
                  model: antideepfake
                  score: 0.0018
                  evidence: []
        '400':
          $ref: '#/components/responses/InvalidAudioFile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/DetectionServiceUnavailable'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/analysis/video:
    post:
      tags: [Analysis]
      operationId: analyzeVideo
      summary: 영상 AI 생성 여부(얼굴조작 딥페이크) 분석 요청
      description: >-
        업로드한 영상의 얼굴조작(face-swap) 딥페이크 여부 분석을 비동기 작업으로 접수한다.
        처리에 수십 초~수 분이 걸릴 수 있어 즉시 202로 jobId를 반환하고, 결과는
        GET /api/v1/analysis/jobs/{jobId}로 폴링해 확인한다. 완전 생성형(Sora류) 영상 탐지는
        지원하지 않는다 - 얼굴조작 딥페이크만 판독한다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 분석할 영상 파일(mp4/mov/avi, 최대 100MB).
      responses:
        '202':
          description: 분석 작업 접수됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobAcceptedResponse'
        '400':
          $ref: '#/components/responses/InvalidVideoFile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/analysis/jobs/{jobId}:
    get:
      tags: [Analysis]
      operationId: getAnalysisJob
      summary: 분석 작업 상태/결과 조회
      description: >-
        POST /api/v1/analysis/video 로 접수한 작업의 현재 상태를 조회한다. status만으로 각 필드의
        존재 여부를 추측하지 말고 aiDetection/scamDetection은 항상 각각 null 체크할 것.


        경우별 응답:

        - AI 판독과 사기감지 중 하나라도 처리 중 오류가 나면 성공한 쪽 결과도 함께 버리고
          FAILED + ANALYSIS_FAILED가 된다(aiDetection/scamDetection 둘 다 없음). 한쪽 결과만
          담긴 COMPLETED는 오지 않는다 - 사기감지가 실패했는데 결과가 비어 있으면 "사기 아님"으로
          오해될 수 있어서다. 같은 영상으로 다시 시도할 수 있다.

        - 영상에서 얼굴을 찾지 못하면 COMPLETED + NO_FACE_DETECTED이고 aiDetection이 없다.
          사기감지는 얼굴과 무관하게 계산되므로 scamDetection은 있을 수 있다.

        - 얼굴도 없고 분석할 텍스트(말소리·화면 글자)도 없으면 COMPLETED + NO_FACE_DETECTED이고
          aiDetection/scamDetection 둘 다 없다. 오류가 아니라 "판독할 대상이 없었다"는 정상 결과다.

        - 얼굴은 있지만 텍스트가 없으면 COMPLETED이고 errorCode 없이 scamDetection만 없다.

        - errorCode 없이 aiDetection/scamDetection이 둘 다 없는 COMPLETED는 오지 않는다.


        PENDING/PROCESSING 응답에는 Retry-After 헤더(초)가 붙는다 - 다음 조회까지 그만큼 기다리면
        된다. 완료(COMPLETED/FAILED)된 작업은 삭제하지 않으므로(회원 탈퇴 시 제외) 나중에 다시
        조회해도 결과를 받을 수 있다. 본인이 접수한 작업이 아니거나, 이미지/음성 분석 기록의
        id를 넣으면 404를 반환한다.
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 조회 성공
          headers:
            Retry-After:
              description: status가 PENDING/PROCESSING일 때만 붙는다. 다음 조회까지 기다릴 시간(초).
              schema:
                type: integer
                example: 5
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
              examples:
                pending:
                  summary: 대기 중 - 앞에 2건 대기
                  value:
                    jobId: 11111111-1111-1111-1111-111111111111
                    status: PENDING
                    jobsAhead: 2
                processing:
                  summary: 처리 중
                  value:
                    jobId: 11111111-1111-1111-1111-111111111111
                    status: PROCESSING
                completed:
                  summary: 분석 완료
                  value:
                    jobId: 11111111-1111-1111-1111-111111111111
                    status: COMPLETED
                    aiDetection:
                      model: dfdc
                      score: 0.91
                      evidence:
                        - title: 얼굴 조작 의심 구간
                          description: 3.0초~7.0초 구간에서 얼굴 합성 흔적이 감지됨
                          tags: [temporal, face-swap]
                          startSec: 3.0
                          endSec: 7.0
                      evidenceImage: "iVBORw0KGgoAAAANSUhEUgAA... (base64 PNG, 생략)"
                    scamDetection:
                      model: lilju
                      score: 0.82
                      evidence:
                        - sentence: 지금 바로 계좌번호와 비밀번호를 알려주셔야 합니다.
                          score: 0.95
                completed_no_face:
                  summary: 분석 완료 - 얼굴 없음(AI판독은 없지만 사기감지는 살아있음)
                  value:
                    jobId: 11111111-1111-1111-1111-111111111111
                    status: COMPLETED
                    scamDetection:
                      model: lilju
                      score: 0.82
                      evidence:
                        - sentence: 지금 바로 계좌번호와 비밀번호를 알려주셔야 합니다.
                          score: 0.95
                    errorCode: NO_FACE_DETECTED
                    errorMessage: 영상에서 얼굴을 찾을 수 없어 AI판독은 제공되지 않습니다. 얼굴이 잘 보이는 영상이면 판독도 함께 받을 수 있습니다.
                completed_no_face_no_text:
                  summary: 분석 완료 - 얼굴도 없고 텍스트도 없음(판독 결과 둘 다 없음, 오류 아님)
                  value:
                    jobId: 11111111-1111-1111-1111-111111111111
                    status: COMPLETED
                    errorCode: NO_FACE_DETECTED
                    errorMessage: 영상에서 얼굴을 찾을 수 없어 AI판독은 제공되지 않습니다. 얼굴이 잘 보이는 영상이면 판독도 함께 받을 수 있습니다.
                completed_no_text:
                  summary: 분석 완료 - 얼굴은 있지만 텍스트 없음(scamDetection만 없음)
                  value:
                    jobId: 11111111-1111-1111-1111-111111111111
                    status: COMPLETED
                    aiDetection:
                      model: dfdc
                      score: 0.12
                      evidence: []
                failed:
                  summary: 분석 완전 실패
                  value:
                    jobId: 11111111-1111-1111-1111-111111111111
                    status: FAILED
                    errorCode: ANALYSIS_FAILED
                    errorMessage: 영상 분석 중 오류가 발생했습니다.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AnalysisJobNotFound'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/analysis/records:
    get:
      tags: [Analysis]
      operationId: getAnalysisRecords
      summary: 내 분석 기록 목록 조회
      description: >-
        로그인한 회원이 완료한 분석(이미지/음성/영상) 기록을 최신순으로 최대 10건 반환한다.
        페이지네이션은 지원하지 않는다 — 처리중/실패한 기록은 포함되지 않으며, 영상 진행 상태
        확인은 GET /api/v1/analysis/jobs/{jobId}를 사용한다.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisRecordListResponse'
              examples:
                image_record:
                  summary: 이미지 기록 - imageDetection만 채워짐
                  value:
                    content:
                      - id: 11111111-1111-1111-1111-111111111111
                        modality: IMAGE
                        createdAt: '2026-09-23T09:00:00Z'
                        mediaAvailable: true
                        imageDetection:
                          model: spai
                          score: 0.87
                          evidenceImage: "iVBORw0KGgoAAAANSUhEUgAA... (base64 PNG, 생략)"
                audio_record:
                  summary: 음성 기록 - audioDetection만 채워짐
                  value:
                    content:
                      - id: 22222222-2222-2222-2222-222222222222
                        modality: AUDIO
                        createdAt: '2026-09-23T08:30:00Z'
                        mediaAvailable: true
                        audioDetection:
                          model: antideepfake
                          score: 0.73
                          evidence:
                            - title: 합성 음성 의심 구간
                              description: 1.0초~4.0초 구간에서 부자연스러운 음성 합성 흔적이 감지됨
                              tags: [temporal]
                              startSec: 1.0
                              endSec: 4.0
                        scamDetection:
                          model: lilju
                          score: 0.82
                          evidence:
                            - sentence: 지금 바로 계좌번호와 비밀번호를 알려주셔야 합니다.
                              score: 0.95
                video_record_no_face:
                  summary: 영상 기록 - 얼굴 없음(videoDetection은 없지만 사기감지는 살아있음)
                  value:
                    content:
                      - id: 33333333-3333-3333-3333-333333333333
                        modality: VIDEO
                        createdAt: '2026-09-23T08:00:00Z'
                        mediaAvailable: false
                        scamDetection:
                          model: lilju
                          score: 0.82
                          evidence:
                            - sentence: 지금 바로 계좌번호와 비밀번호를 알려주셔야 합니다.
                              score: 0.95
                        errorCode: NO_FACE_DETECTED
                        errorMessage: 영상에서 얼굴을 찾을 수 없어 AI판독은 제공되지 않습니다. 얼굴이 잘 보이는 영상이면 판독도 함께 받을 수 있습니다.
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/analysis/records/{id}/media:
    get:
      tags: [Analysis]
      operationId: getAnalysisRecordMedia
      summary: 분석 원본 파일 다운로드
      description: >-
        분석 기록의 원본 파일(업로드한 이미지/음성/영상)을 업로드 때의 형식 그대로 내려준다.
        Range 요청을 지원한다(Range: bytes=시작-끝) - 영상을 끝까지 받기 전에 재생·탐색할 수 있다.


        보관 정책:

        - 회원당 최신 10건의 원본만 보관한다(기록 목록과 같은 기준). 11번째부터는 원본만 삭제되고
          분석 결과는 남는다.

        - 분석이 완료된 기록만 원본을 보관한다(처리 중/실패한 영상은 원본 없음). 이 기능이 생기기
          전의 기록도 원본이 없다.

        - 원본은 저장소에서 암호화된 상태로 보관되며, 회원 탈퇴 시 전부 삭제된다.


        원본이 없으면 404 MEDIA_NOT_AVAILABLE을 반환한다. 기록 목록의 mediaAvailable이 false인
        기록은 요청하지 않아도 된다.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 분석 기록 id(기록 목록의 id, 이미지/음성 분석 응답의 id, 영상의 jobId와 같은 값).
          schema:
            type: string
            format: uuid
        - name: Range
          in: header
          required: false
          description: 일부만 받을 때 사용. 예) bytes=0-1048575
          schema:
            type: string
      responses:
        '200':
          description: 원본 파일 전체. Content-Type은 업로드 형식 그대로, Content-Length 포함.
          headers:
            Content-Length:
              schema:
                type: integer
                format: int64
            Accept-Ranges:
              schema:
                type: string
                example: bytes
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '206':
          description: Range 요청에 대한 일부 응답. Content-Range로 범위를 알려준다.
          headers:
            Content-Length:
              schema:
                type: integer
                format: int64
            Content-Range:
              schema:
                type: string
                example: bytes 0-1048575/52428800
            Accept-Ranges:
              schema:
                type: string
                example: bytes
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/MediaNotFound'
        '416':
          $ref: '#/components/responses/RangeNotSatisfiable'
        default:
          $ref: '#/components/responses/ServerError'
  /api/v1/analysis/report:
    get:
      tags: [Analysis]
      operationId: getAnalysisReport
      summary: 내 분석 리포트(통계) 조회
      description: >-
        로그인한 회원의 완료된 분석 기록을 집계한 통계를 반환한다. aiDetectedCount/scamDetectedCount는
        점수 0.5 이상을 "탐지됨"으로 판단한 건수다.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT 액세스 토큰. `Authorization: Bearer <accessToken>` 헤더로 전달한다."
  schemas:
...
    ImageAnalysisResponse:
      type: object
      required: [id, aiDetection]
      properties:
        id:
          type: string
          format: uuid
          nullable: false
          description: >-
            이번 분석으로 생성된 분석 기록 id. 기록 목록(GET /api/v1/analysis/records)의 id와 같은
            값이며, 원본 다운로드(GET /api/v1/analysis/records/{id}/media)에 쓴다. 영상의 jobId와
            같은 역할이지만 폴링용이 아니다(GET /api/v1/analysis/jobs/{jobId}는 영상 전용).
          example: 11111111-1111-1111-1111-111111111111
        aiDetection:
          $ref: '#/components/schemas/ImageDetectionResult'
        scamDetection:
          allOf:
            - $ref: '#/components/schemas/ScamDetectionResult'
          nullable: true
          description: >-
            사기(보이스피싱 등) 위험도 판독 결과. 텍스트가 전혀 추출되지 않으면 없음(null) - 이 필드가
            null이면 항상 "분석할 텍스트가 없었다"는 뜻이다. 사기감지 파이프라인 자체가 실패한 경우는
            null로 조용히 넘어가지 않고 502로 요청 전체가 실패한다(2026-09-22).
    AudioAnalysisResponse:
      type: object
      required: [id, aiDetection]
      properties:
        id:
          type: string
          format: uuid
          nullable: false
          description: >-
            이번 분석으로 생성된 분석 기록 id. 기록 목록(GET /api/v1/analysis/records)의 id와 같은
            값이며, 원본 다운로드(GET /api/v1/analysis/records/{id}/media)에 쓴다. 영상의 jobId와
            같은 역할이지만 폴링용이 아니다(GET /api/v1/analysis/jobs/{jobId}는 영상 전용).
          example: 22222222-2222-2222-2222-222222222222
        aiDetection:
          $ref: '#/components/schemas/AudioDetectionResult'
        scamDetection:
          allOf:
            - $ref: '#/components/schemas/ScamDetectionResult'
          nullable: true
          description: >-
            사기(보이스피싱 등) 위험도 판독 결과. 텍스트가 전혀 추출되지 않으면 없음(null) - 이 필드가
            null이면 항상 "분석할 텍스트가 없었다"는 뜻이다. 사기감지 파이프라인 자체가 실패한 경우는
            null로 조용히 넘어가지 않고 502로 요청 전체가 실패한다(2026-09-22).
    AnalysisJobAcceptedResponse:
      type: object
      required: [jobId]
      properties:
        jobId:
          type: string
          format: uuid
          nullable: false
          description: 생성된 분석 작업 ID. GET /api/v1/analysis/jobs/{jobId}로 상태를 조회한다.
          example: 11111111-1111-1111-1111-111111111111
    AnalysisJobResponse:
      type: object
      required: [jobId, status]
      properties:
        jobId:
          type: string
          format: uuid
          nullable: false
          example: 11111111-1111-1111-1111-111111111111
        status:
          type: string
          nullable: false
          enum: [PENDING, PROCESSING, COMPLETED, FAILED]
          description: 작업 처리 상태.
          example: COMPLETED
        jobsAhead:
          type: integer
          format: int32
          nullable: true
          minimum: 0
          description: >-
            status가 PENDING일 때만 채워진다. 이 작업보다 먼저 접수돼 아직 처리를 시작하지 않은
            영상 작업 수("앞에 N건 대기 중" 표시용). 영상은 접수 순서대로 처리된다. 0이어도
            PENDING일 수 있다(처리 중인 작업이 끝나기를 기다리는 중).
          example: 2
        aiDetection:
          allOf:
            - $ref: '#/components/schemas/VideoDetectionResult'
          nullable: true
          description: >-
            status가 FAILED이면 항상 없다. status가 COMPLETED여도 영상에서 얼굴을 찾지 못하면
            없을 수 있다(그 경우 errorCode가 NO_FACE_DETECTED로 채워짐) - status만 보고 존재를
            추측하지 말고 항상 null 체크할 것.
        scamDetection:
          allOf:
            - $ref: '#/components/schemas/ScamDetectionResult'
          nullable: true
          description: >-
            status가 COMPLETED이고 분석할 텍스트가 있었을 때만 채워진다. 영상의 텍스트는 말소리(음성
            인식)와 화면 글자(자막 등)를 합친 것이다 - 이 필드가 없으면 "말소리도 화면 글자도
            없었다"는 뜻이다(말소리만 없고 화면 글자가 있으면 채워진다). 사기감지 처리 자체가 실패한
            경우는 이 필드가 조용히 비지 않고 작업 전체가 FAILED가 된다. aiDetection 유무와는 무관하다.
        errorCode:
          type: string
          nullable: true
          description: >-
            채워지는 두 경우가 있다: (1) status가 FAILED일 때 - 완전 실패 사유. (2) status가
            COMPLETED인데 일부 판독만 정상적으로 비어있을 때 - 부분 사유. 현재 값: NO_FACE_DETECTED(영상에서
            얼굴을 찾지 못해 aiDetection만 비어있음 - COMPLETED와 함께 옴, scamDetection은 있을 수 있음),
            ANALYSIS_FAILED(AI 판독이나 사기감지 중 하나라도 처리 중 오류 - FAILED와 함께 옴, 같은 영상으로
            재시도 가능). errorMessage와
            달리 문구가 바뀌어도 값은 유지되므로 분기 처리엔 이 필드를 쓴다.
          example: NO_FACE_DETECTED
        errorMessage:
          type: string
          nullable: true
          description: errorCode와 같은 경우에 채워지는, 사용자에게 그대로 보여줄 문구. 분기 처리에는 errorCode를 쓴다.
    AnalysisRecordListResponse:
      type: object
      required: [content]
      description: 최신순 최대 10건. 페이지네이션 없음.
      properties:
        content:
          type: array
          nullable: false
          maxItems: 10
          items:
            $ref: '#/components/schemas/AnalysisRecordSummary'
...
    AnalysisReportResponse:
      type: object
      required: [totalCount, imageCount, audioCount, videoCount, aiDetectedCount, scamDetectedCount]
      description: >-
        완료된 분석 기록 기준 집계. aiDetectedCount/scamDetectedCount는 점수 0.5 이상을
        "탐지됨"으로 판단한 건수다(고정 임계값, 추후 조정 가능).
      properties:
        totalCount:
          type: integer
          format: int64
          nullable: false
          example: 23
        imageCount:
          type: integer
          format: int64
          nullable: false
          example: 10
        audioCount:
          type: integer
          format: int64
          nullable: false
          example: 8
        videoCount:
          type: integer
          format: int64
          nullable: false
          example: 5
        aiDetectedCount:
          type: integer
          format: int64
          nullable: false
          description: AI 생성물로 판별된(aiScore >= 0.5) 건수.
          example: 3
        scamDetectedCount:
          type: integer
          format: int64
          nullable: false
          description: 사기 위험으로 판별된(scamScore >= 0.5) 건수.
          example: 2
...
    ImageDetectionResult:
      type: object
      required: [model, score]
      description: >-
        이미지 AI 판독 결과. SPAI는 근거 카드(evidence)를 만들지 않아 그 필드는 없다.
        evidenceImage는 SPAI의 attention 히트맵(원본 사진 위에 합성된 오버레이)이다.
      properties:
        model:
          type: string
          nullable: false
          description: 탐지에 사용된 모델 이름.
          example: spai
        score:
          type: number
          format: double
          nullable: false
          minimum: 0
          maximum: 1
          description: AI 생성물일 확률(0~1). 1에 가까울수록 AI 생성 가능성이 높다.
          example: 0.000134
        evidenceImage:
          type: string
          nullable: true
          description: >-
            판독 근거 히트맵 이미지(base64 인코딩 PNG). 이미지의 어느 부분이 의심스러운지
            시각적으로 보여준다(best-effort - 실패하면 null).
    AudioDetectionResult:
      type: object
      required: [model, score, evidence]
      description: 음성 AI 판독 결과. 공간적 근거(이미지) 개념이 없어 evidenceImage 필드는 없다.
      properties:
        model:
          type: string
          nullable: false
          description: 탐지에 사용된 모델 이름.
          example: antideepfake
        score:
          type: number
          format: double
          nullable: false
          minimum: 0
          maximum: 1
          description: AI 생성물일 확률(0~1). 1에 가까울수록 AI 생성 가능성이 높다.
          example: 0.0018
        evidence:
          type: array
          nullable: false
          description: 판독 근거 카드 목록. 근거가 없으면 빈 배열.
          items:
            $ref: '#/components/schemas/Evidence'
    VideoDetectionResult:
      type: object
      required: [model, score, evidence]
      description: 영상 AI 판독 결과. 시간 구간 근거와 공간적 히트맵을 둘 다 지원하는 유일한 포맷.
      properties:
        model:
          type: string
          nullable: false
          description: 탐지에 사용된 모델 이름.
          example: dfdc
        score:
          type: number
          format: double
          nullable: false
          minimum: 0
          maximum: 1
          description: AI 생성물일 확률(0~1). 1에 가까울수록 AI 생성 가능성이 높다.
          example: 0.91
        evidence:
          type: array
          nullable: false
          description: 판독 근거 카드 목록. 근거가 없으면 빈 배열.
          items:
            $ref: '#/components/schemas/Evidence'
        evidenceImage:
          type: string
          nullable: true
          description: >-
            판독 근거 히트맵 이미지(base64 인코딩 PNG). 얼굴의 어느 부분이 의심스러운지
            시각적으로 보여준다(best-effort - 실패하면 null).
    ScamDetectionResult:
      type: object
      required: [model, score, evidence]
      description: 사기(보이스피싱 등) 위험도 판독 결과. 텍스트가 전혀 추출되지 않으면 이 필드 자체가 없다(null).
      properties:
        model:
          type: string
          nullable: false
          description: 사기감지에 사용된 모델 이름.
          example: lilju
        score:
          type: number
          format: double
          nullable: false
          minimum: 0
          maximum: 1
          description: 사기(보이스피싱 등)일 위험도(0~1). 문장별 확률의 최댓값x0.7 + 평균x0.3으로 집계.
          example: 0.82
        evidence:
          type: array
          nullable: false
          description: 위험하다고 판단된 문장 근거 카드 목록(개별 확률 0.5 이상만 포함). 근거가 없으면 빈 배열.
          items:
            $ref: '#/components/schemas/ScamEvidence'
...
  responses:
...
    InvalidImageFile:
      description: 업로드한 이미지 파일이 비어있거나 지원하지 않는 형식/용량임
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://api.veritae.app/errors/invalid-image-file
            title: Invalid Image File
            status: 400
            detail: '지원하지 않는 파일 형식입니다: text/plain'
            instance: /api/v1/analysis/image
            errorCode: INVALID_IMAGE_FILE
            timestamp: '2026-08-12T09:00:00Z'
    InvalidAudioFile:
      description: >-
        업로드한 음성 파일이 비어있거나 지원하지 않는 형식/용량임(INVALID_AUDIO_FILE), 또는 재생 길이가
        5분을 초과함(AUDIO_TOO_LONG)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            invalid_file:
              summary: 형식/용량 오류
              value:
                type: https://api.veritae.app/errors/invalid-audio-file
                title: Invalid Audio File
                status: 400
                detail: 파일 용량이 25MB를 초과합니다.
                instance: /api/v1/analysis/audio
                errorCode: INVALID_AUDIO_FILE
                timestamp: '2026-08-19T09:00:00Z'
            too_long:
              summary: 5분 초과
              value:
                type: https://api.veritae.app/errors/audio-too-long
                title: Audio Too Long
                status: 400
                detail: 음성 길이가 5분을 초과합니다.
                instance: /api/v1/analysis/audio
                errorCode: AUDIO_TOO_LONG
                timestamp: '2026-09-26T09:00:00Z'
    InvalidVideoFile:
      description: 업로드한 영상 파일이 비어있거나 지원하지 않는 형식/용량임
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://api.veritae.app/errors/invalid-video-file
            title: Invalid Video File
            status: 400
            detail: 파일 용량이 100MB를 초과합니다.
            instance: /api/v1/analysis/video
            errorCode: INVALID_VIDEO_FILE
            timestamp: '2026-08-26T09:00:00Z'
...
    Unauthorized:
      description: 인증 실패(액세스 토큰 누락/무효/만료)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://api.veritae.app/errors/unauthorized
            title: Unauthorized
            status: 401
            detail: 인증이 필요합니다.
            instance: /api/v1/members/me
            errorCode: UNAUTHORIZED
            timestamp: '2026-07-17T09:00:00Z'
    DetectionServiceUnavailable:
      description: 탐지 서버(별도 인프라) 호출 실패(네트워크 오류, 타임아웃, 5xx 등)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://api.veritae.app/errors/detection-service-unavailable
            title: Detection Service Unavailable
            status: 502
            detail: 탐지 서버 호출에 실패했습니다.
            instance: /api/v1/analysis/image
            errorCode: DETECTION_SERVICE_UNAVAILABLE
            timestamp: '2026-08-12T09:00:00Z'
    AnalysisJobNotFound:
      description: 분석 작업을 찾을 수 없거나 본인 소유가 아님
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://api.veritae.app/errors/analysis-job-not-found
            title: Analysis Job Not Found
            status: 404
            detail: 분석 작업을 찾을 수 없습니다.
            instance: /api/v1/analysis/jobs/11111111-1111-1111-1111-111111111111
            errorCode: ANALYSIS_JOB_NOT_FOUND
            timestamp: '2026-08-26T09:00:00Z'
    MediaNotFound:
      description: >-
        기록이 없거나 본인 소유가 아님(ANALYSIS_RECORD_NOT_FOUND), 또는 기록은 있지만 원본이
        없음(MEDIA_NOT_AVAILABLE - 보관 기간 지남, 보관 기능 이전 기록 등)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            record_not_found:
              summary: 기록 없음/본인 소유 아님
              value:
                type: https://api.veritae.app/errors/analysis-record-not-found
                title: Analysis Record Not Found
                status: 404
                detail: 분석 기록을 찾을 수 없습니다.
                instance: /api/v1/analysis/records/11111111-1111-1111-1111-111111111111/media
                errorCode: ANALYSIS_RECORD_NOT_FOUND
                timestamp: '2026-09-26T09:00:00Z'
            media_not_available:
              summary: 원본 없음
              value:
                type: https://api.veritae.app/errors/media-not-available
                title: Media Not Available
                status: 404
                detail: 이 기록의 원본 파일이 없습니다.
                instance: /api/v1/analysis/records/11111111-1111-1111-1111-111111111111/media
                errorCode: MEDIA_NOT_AVAILABLE
                timestamp: '2026-09-26T09:00:00Z'
    RangeNotSatisfiable:
      description: Range 요청의 범위가 파일 크기를 벗어남
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://api.veritae.app/errors/range-not-satisfiable
            title: Range Not Satisfiable
            status: 416
            detail: 요청한 범위가 파일 크기를 벗어납니다.
            instance: /api/v1/analysis/records/11111111-1111-1111-1111-111111111111/media
            errorCode: RANGE_NOT_SATISFIABLE
            timestamp: '2026-09-26T09:00:00Z'
...
    ServerError:
      description: 예상하지 못한 서버 오류(예 400/401/409 외 나머지 모든 상태코드)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://api.veritae.app/errors/internal-server-error
            title: Internal Server Error
            status: 500
            detail: 서버에서 예상하지 못한 오류가 발생했습니다.
            instance: /api/v1/auth/login
            errorCode: INTERNAL_SERVER_ERROR
            timestamp: '2026-07-17T09:00:00Z'